% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheCall.R
\name{cacheCall}
\alias{cacheCall}
\title{Call to Cache}
\usage{
cacheCall(fnName, args, pipeName = NULL, cachePath = "", ...)
}
\arguments{
\item{fnName}{the name of the function (any function) being called.}

\item{args}{a list of arguments that will define the filename.}

\item{pipeName}{Can be the name of the parent function or the name of the pipeline if there is one. Can also simply be an ID for a series of functions and files that should be grouped together. Defaults to \code{NULL}.}

\item{cachePath}{a character string providing path to the Cache directory.}

\item{...}{arguments that will be passed to fnName call if it is executed.}
}
\value{
the output of the function call, either from a read-in file or from an executed call. If the latter, the corresponding file will also be saved in \code{cachePath}.
}
\description{
This function acts as a wrapper, such that for every real function call, enclosing it with \code{callCache()} first checks whether the output of the function call exists in \code{cachePath}.
If the file does not exist, it executes the function call and saves the output to \code{cachePath}. If the file does exist, it reads the file.
The filename is generated here and the file naming convention is determined by the function name(s) and arguments provided.
}
